package/test-utils: move some test utils to separate package so test-utils can depend on core
