+++ packages/core/package.json
@@ -42,7 +42,7 @@
   },
   "devDependencies": {
     "@backstage/cli": "^0.1.1-alpha.3",
-    "@backstage/test-utils": "^0.1.1-alpha.3",
+    "@backstage/test-utils-core": "^0.1.1-alpha.3",
     "@backstage/theme": "^0.1.1-alpha.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
