+++ packages/test-utils-core/.npmrc
@@ -1,0 +1,1 @@
+registry=https://registry.npmjs.org/
