+++ packages/test-utils-core/src/index.ts
@@ -14,4 +14,3 @@
  * limitations under the License.
  */
 export * from './testUtils';
-export * from '@backstage/test-utils-core';
