+++ packages/test-utils-core/src/setupTests.ts
@@ -13,5 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './testUtils';
-export * from '@backstage/test-utils-core';
+
+import '@testing-library/jest-dom/extend-expect';
