+++ packages/test-utils/package.json
@@ -39,6 +39,7 @@
   },
   "peerDependencies": {
     "@backstage/theme": "^0.1.1-alpha.3",
+    "@backstage/test-utils-core": "^0.1.1-alpha.3",
     "@material-ui/core": "^4.9.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
